Do not fail if some turngames can't be loaded
In the event of some turngame referenced by the coordinator aren't
found, triggering a failure is a bad idea: the failure should only
be triggered if NO games can be loaded.

Only in case of success the course of actions can continue (and end
up with refreshing the screen). We don't want that to be interupted
for a few games that failed to load.