Extract Flash Ajax impl in own class
While AIR SDK supports RESTful APIs, it's not the case with
the flash player. To solve this, I successfully tested using
the JS Ajax class from the browser. A solution would then be to
implement a Flash Ajax proxy that bridges to the JS Ajax impl.

This commit is the first step, it allows the backend flash Ajax
implementation to be selected at runtime, by setting a custom
`Ajax.implFactory`, that builds the desired IAjax implementation.